Move all newick stuff from #170 to their new package.
