+++ src/main/java/nl/tudelft/context/model/newick/Node.java
@@ -3,8 +3,8 @@ package nl.tudelft.context.model.newick;
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import nl.tudelft.context.drawable.DrawableNode;
-import nl.tudelft.context.newick.selection.None;
-import nl.tudelft.context.newick.selection.Selection;
+import nl.tudelft.context.model.newick.selection.None;
+import nl.tudelft.context.model.newick.selection.Selection;
 
 import java.util.ArrayList;
 import java.util.HashSet;
