+++ src/main/java/nl/tudelft/context/model/newick/selection/None.java
@@ -1,25 +1,25 @@
-package nl.tudelft.context.newick.selection;
+package nl.tudelft.context.model.newick.selection;
 
 /**
  * @author René Vennik <renevennik@gmail.com>
  * @version 1.0
  * @since 22-5-2015
  */
-public class All implements Selection {
+public class None implements Selection {
 
     @Override
     public boolean useSources() {
-        return true;
+        return false;
     }
 
     @Override
     public Selection toggle() {
-        return new None();
+        return new All();
     }
 
     @Override
     public Selection merge(final Selection selection) {
-        if (selection instanceof All) {
+        if (selection instanceof None) {
             return this;
         }
         return new Partial();
@@ -27,7 +27,7 @@ public class All implements Selection {
 
     @Override
     public String styleClass() {
-        return "selected";
+        return "none";
     }
 
 }
