+++ src/main/java/nl/tudelft/context/model/newick/selection/package-info.java
@@ -5,4 +5,4 @@
  * @version 1.0
  * @since 22-5-2015
  */
-package nl.tudelft.context.newick.selection;
+package nl.tudelft.context.model.newick.selection;
